Module5-Model_Deployment/requirements.txt
pillow==10.0.1
flask==3.0.3
numpy==1.23.5
pandas==2.0.3
tensorflow==2.15.0
h5py==3.10.0